+++ tests/src/all_suites.adb
@@ -2,6 +2,7 @@ with Proto_Tests;
 with Foundation_Tests;
 with Result_Tests;
 with Context_Tests;
+with Engine_Tests;
 
 package body All_Suites is
 
@@ -13,6 +14,7 @@ package body All_Suites is
       Foundation_Tests.Add_Tests (Result);
       Result_Tests.Add_Tests (Result);
       Context_Tests.Add_Tests (Result);
+      Engine_Tests.Add_Tests (Result);
       return Result;
    end Suite;